+++ TrafficCapture/dockerSolution/src/main/docker/migrationConsole/lib/integ_test/integ_test/test_cases/ma_test_base.py
@@ -1,8 +1,6 @@
 from enum import Enum
 
-import requests
-
-from ..common_utils import wait_for_service_status
+from ..common_utils import check_ma_system_health, wait_for_service_status
 from ..cluster_version import ClusterVersion, is_incoming_version_supported
 from ..operations_library_factory import get_operations_library_by_version
 
@@ -22,17 +20,6 @@ class ClusterVersionCombinationUnsupported(Exception):
         self.target_version = target_version
         self.message = f"{message}: Source version '{source_version}' and Target version '{target_version}'"
         super().__init__(self.message)
-
-
-def check_ma_system_health():
-    print("check_ma_system_health")
-    resp = requests.get("http://127.0.0.1:80/api/system/health")
-    assert resp.status_code == 200
-    data = resp.json()
-    assert data["status"] == "ok"
-    assert "checks" in data
-    assert all(val == "ok" for val in data["checks"].values())
-    print("check_ma_system_health complete")
 
 
 class MATestBase:
