+++ TrafficCapture/dockerSolution/src/main/docker/migrationConsole/start-console.sh
@@ -11,16 +11,9 @@ mkdir -p "$LOG_DIR"
 
 cd /root/lib/console_link
 export FASTAPI_ROOT_PATH=/api
-pipenv run gunicorn console_link.api.main:app \
+exec pipenv run gunicorn console_link.api.main:app \
     -k uvicorn.workers.UvicornWorker \
     -w 4 \
     -b 0.0.0.0:8000 \
     --access-logfile "$LOG_DIR/access.log" \
-    --error-logfile "$LOG_DIR/error.log" &
-
-sleep 2
-
-cd /root
-
-# Keep container running and interactive
-exec tail -f /dev/null
+    --error-logfile "$LOG_DIR/error.log"
